Add files via upload
Here is the admin page. You need to add adjust the connectToDB.inc file for it to work and make sure that the admin page and the connectToDB.inc file are in the same folder. If there are small, syntax errors, I'm really sorry. One thing that's really troublesome with web development is that it's really picky. One stray letter or small mistake, and it does not work.